pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.4",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "playwright>=1.54.0",
    "requests>=2.32.4",
    "trafilatura>=2.0.0",
]
